Setup webpack, eslint, start to work with exports. In progress.
