+++ public/index.html
@@ -85,7 +85,7 @@
       </section>
 
       <section class="main__filter filter container">
-        <input
+        <!-- <input
           type="radio"
           id="filter__all"
           class="filter__input visually-hidden"
@@ -150,7 +150,7 @@
         />
         <label for="filter__archive" class="filter__label"
           >ARCHIVE <span class="filter__archive-count">115</span></label
-        >
+        > -->
       </section>
 
       <section class="statistic container visually-hidden">
@@ -205,7 +205,7 @@
         </p>
 
         <div class="board__tasks">
-          <article class="card card--edit card--black">
+          <!-- <article class="card card--edit card--black">
             <form class="card__form" method="get">
               <div class="card__inner">
                 <div class="card__control">
@@ -1916,13 +1916,13 @@ Here is a card with filled data</textarea
                 </div>
               </div>
             </form>
-          </article>
+          </article> -->
         </div>
 
         <button class="load-more" type="button">load more</button>
       </section>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
