+++ src/filters.js
@@ -1,0 +1,33 @@
+export const filters = [
+  {
+    name: `ALL`,
+    tasksAmount: 32,
+    checked: true
+  },
+  {
+    name: `OVERDUE`,
+    tasksAmount: 1,
+    disabled: true
+  },
+  {
+    name: `TODAY`,
+    tasksAmount: 2,
+    disabled: true
+  },
+  {
+    name: `FAVORITES`,
+    tasksAmount: 3
+  },
+  {
+    name: `Repeating`,
+    tasksAmount: 5
+  },
+  {
+    name: `Tags`,
+    tasksAmount: 8
+  },
+  {
+    name: `ARCHIVE`,
+    tasksAmount: 13
+  }
+];
