+++ webpack.config.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+module.exports = {
+  mode: 'development',
+  entry: './src/main.js',
+  devtool: 'inline-source-map',
+  output: {
+    filename: 'bundle.js',
+    path: path.join(__dirname, 'public')
+  }
+};
